Replace email auth with key auth
Auth previously worked by using access control lists based on
email identities. To gain access the user would verify their
email address which would provide the client with a token. This
required the ACLs to be set up ahead of time, and made GemDrive
depend on SMTP and all the pain that comes with email.

This commit starts moving towards a simpler auth model where a
user can use a key to generate other keys with equal or lesser
privileges. This creates a tree of keys, with a root key at the
top which is generated when a fresh database is created.