Remove built-in login page
Used to return a staic HTML page along with 403s. Was more useful
when GemDrive supported cookies, as it allowed you to log in once
on each device and then be able to open images and other files
directly in the browser. However cookies seem to present a lot of
security challenges, so they're removed for now. So now 403s just
need to return a simple text error.